harder review comments: moved the get_extra_links to serialisedBaseOperator, some refactoring, signature of GOL to BOL
